requirements.txt
requests
configparser
jupyter
pandas
numpy
python-docx
lxml
bs4